+++ README.md
@@ -6,6 +6,7 @@ Ecommerce项目包括：
 |代码库|用途|地址|
 | --- | --- | --- |
 |ecommerce-order-service|Order服务|[https://github.com/e-commerce-sample/ecommerce-order-service](https://github.com/e-commerce-sample/ecommerce-order-service)|
+|ecommerce-order-query-service|Order查询服务|[https://github.com/e-commerce-sample/ecommerce-order-query-service](https://github.com/e-commerce-sample/ecommerce-order-query-service)|
 |ecommerce-product-service|Product服务|[https://github.com/e-commerce-sample/ecommerce-product-service](https://github.com/e-commerce-sample/ecommerce-product-service)|
 |ecommerce-inventory-service|Inventory服务|[https://github.com/e-commerce-sample/ecommerce-inventory-service](https://github.com/e-commerce-sample/ecommerce-inventory-service)|
 |ecommerce-shared-model|共享模型，不含Spring|[https://github.com/e-commerce-sample/ecommerce-shared-model](https://github.com/e-commerce-sample/ecommerce-shared-model)|
