+++ ecommerce-order-service-sdk/src/main/java/com/ecommerce/order/sdk/representation/order/OrderItem.java
@@ -1,27 +1,12 @@
 package com.ecommerce.order.sdk.representation.order;
 
+import lombok.Value;
+
 import java.math.BigDecimal;
 
+@Value
 public class OrderItem {
     private final String productId;
     private final int count;
     private final BigDecimal itemPrice;
-
-    public OrderItem(String productId, int count, BigDecimal itemPrice) {
-        this.productId = productId;
-        this.count = count;
-        this.itemPrice = itemPrice;
-    }
-
-    public String getProductId() {
-        return productId;
-    }
-
-    public int getCount() {
-        return count;
-    }
-
-    public BigDecimal getItemPrice() {
-        return itemPrice;
-    }
 }
