+++ local-build.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
-#Use fresh mysql
-./mysql-clean-local.sh
+echo "Stop existing docker containers and remove all its data volumes"
+./gradlew composeDown
 
 #Start build process...
 ./gradlew clean build
